[HttpKernel] Deprecate some compiler passes in favor of tagged iterator args
