+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/ConfigCachePass.php
@@ -13,12 +13,12 @@ namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\Config\DependencyInjection\ConfigCachePass as BaseConfigCachePass;
 
-@trigger_error(sprintf('The %s class is deprecated since version 3.3 and will be removed in 4.0. Use Symfony\Component\Config\DependencyInjection\ConfigCachePass instead.', ConfigCachePass::class), E_USER_DEPRECATED);
+@trigger_error(sprintf('The %s class is deprecated since version 3.3 and will be removed in 4.0. Use tagged iterator arguments instead.', ConfigCachePass::class), E_USER_DEPRECATED);
 
 /**
  * Adds services tagged config_cache.resource_checker to the config_cache_factory service, ordering them by priority.
  *
- * @deprecated since version 3.3, to be removed in 4.0. Use {@link BaseConfigCachePass} instead.
+ * @deprecated since version 3.3, to be removed in 4.0. Use tagged iterator arguments instead.
  *
  * @author Matthias Pigulla <mp@webfactory.de>
  * @author Benjamin Klotz <bk@webfactory.de>
