+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/services.xml
@@ -24,7 +24,7 @@
         <service id="Symfony\Component\HttpFoundation\RequestStack" alias="request_stack" />
 
         <service id="cache_warmer" class="Symfony\Component\HttpKernel\CacheWarmer\CacheWarmerAggregate" public="true">
-            <argument type="collection" />
+            <argument type="tagged" tag="kernel.cache_warmer" />
         </service>
 
         <service id="kernel.class_cache.cache_warmer" class="Symfony\Bundle\FrameworkBundle\CacheWarmer\ClassCacheCacheWarmer">
@@ -40,7 +40,7 @@
         </service>
 
         <service id="cache_clearer" class="Symfony\Component\HttpKernel\CacheClearer\ChainCacheClearer" public="true">
-            <argument type="collection" />
+            <argument type="tagged" tag="kernel.cache_clearer" />
         </service>
 
         <service id="kernel" synthetic="true" public="true" />
@@ -63,7 +63,7 @@
         </service>
 
         <service id="config_cache_factory" class="Symfony\Component\Config\ResourceCheckerConfigCacheFactory">
-            <argument /> <!-- resource checkers -->
+            <argument type="tagged" tag="config_cache.resource_checker" />
         </service>
 
         <service id="Symfony\Component\DependencyInjection\Config\ContainerParametersResourceChecker">
