+++ src/Symfony/Component/HttpKernel/CacheWarmer/CacheWarmerAggregate.php
@@ -15,17 +15,21 @@ namespace Symfony\Component\HttpKernel\CacheWarmer;
  * Aggregates several cache warmers into a single one.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since version 3.4
  */
 class CacheWarmerAggregate implements CacheWarmerInterface
 {
     protected $warmers = array();
     protected $optionalsEnabled = false;
+    private $triggerDeprecation = false;
 
-    public function __construct(array $warmers = array())
+    public function __construct($warmers = array())
     {
         foreach ($warmers as $warmer) {
             $this->add($warmer);
         }
+        $this->triggerDeprecation = true;
     }
 
     public function enableOptionalWarmers()
@@ -53,6 +57,8 @@ class CacheWarmerAggregate implements CacheWarmerInterface
      * Checks whether this warmer is optional or not.
      *
      * @return bool always false
+     *
+     * @deprecated since version 3.4, to be removed in 4.0, inject the list of clearers as a constructor argument instead.
      */
     public function isOptional()
     {
@@ -61,14 +67,23 @@ class CacheWarmerAggregate implements CacheWarmerInterface
 
     public function setWarmers(array $warmers)
     {
+        @trigger_error(sprintf('The "%s()" method is deprecated since version 3.4 and will be removed in 4.0, inject the list of clearers as a constructor argument instead.', __METHOD__), E_USER_DEPRECATED);
+
         $this->warmers = array();
         foreach ($warmers as $warmer) {
             $this->add($warmer);
         }
     }
 
+    /**
+     * @deprecated since version 3.4, to be removed in 4.0, inject the list of clearers as a constructor argument instead.
+     */
     public function add(CacheWarmerInterface $warmer)
     {
+        if ($this->triggerDeprecation) {
+            @trigger_error(sprintf('The "%s()" method is deprecated since version 3.4 and will be removed in 4.0, inject the list of clearers as a constructor argument instead.', __METHOD__), E_USER_DEPRECATED);
+        }
+
         $this->warmers[] = $warmer;
     }
 }
